backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# PyTorch and ML libraries
torch>=2.2.0
transformers==4.35.0
accelerate==0.24.1
protobuf==4.25.0

# Additional utilities
aiofiles==23.2.1
httpx==0.25.1
